+++ docs/www.fpart.org/docs/changelog.md
@@ -9,9 +9,10 @@
       fit in regular partitions when option '-s' *and* non-live mode are used.
       Note for 3rd-party tools: that change may impact you!
       (partly fixes GH issue #36)
-    - fpart: avoid returning an additional -empty- partition when only special
-      partition '0' has been used, such as in:
-      $ echo "abcdef" > /tmp/abcd ; fpart -s 3 /tmp/abcd
+    - fpart: avoid returning an additional -empty- partition '1' when only
+      special partition '0' has been used
+    - fpart: avoid returning special partition '0' when it is empty
+      (partly fixes GH issue #36)
     - fpsync: add more checks for incompatible rsync options
     - fpsync: force non-recursive rsync(1) in file-based mode
 ```